+++ changelog_entry.yaml
@@ -1,4 +1,4 @@
 - bump: minor
   changes:
     added:
-    - Quebec Work Premium Tax Credit.
+    - Quebec work premium tax credit.
