+++ policyengine_canada/parameters/gov/provinces/qc/tax/income/credits/work_premium/adapted_work_premium/couple/amount/with_children.yaml
@@ -3,7 +3,7 @@ values:
   2022-01-01: 4_523.20
 metadata:
   unit: currency-CAD
-  label: Quebec maximum anual adapted work premium amount for couple with at least one child
+  label: Quebec maximum annual adapted work premium amount for couple with at least one child
   reference:
     - title: Revenu Quebec - Amount of the Work Premium Tax Credit
       href: https://www.revenuquebec.ca/en/citizens/tax-credits/work-premium-tax-credits/tax-credit-amounts/
