+++ policyengine_canada/parameters/gov/provinces/qc/tax/income/credits/work_premium/adapted_work_premium/single/amount/single_parent.yaml
@@ -3,7 +3,7 @@ values:
   2022-01-01: 3_628.50
 metadata:
   unit: currency-CAD
-  label: Quebec maximum adapted anual work premium amount for single parent
+  label: Quebec maximum adapted anual work premium amount for single parents
   reference:
     - title: Revenu Quebec - Amount of the Work Premium Tax Credit
       href: https://www.revenuquebec.ca/en/citizens/tax-credits/work-premium-tax-credits/tax-credit-amounts/
