+++ policyengine_canada/tests/gov/provinces/qc/tax/income/credits/work_premium/adapted_work_premium/qc_adapted_work_premium_couple_amount.yaml
@@ -1,0 +1,67 @@
+- name: Quebec adapted work premium amount for couple with children with family income below the turndown threshold
+  period: 2022
+  input:
+    qc_adapted_work_premium_eligibility: true
+    is_married: true
+    count_children: 1
+    adjusted_family_net_income: 20_000
+    province_code: QC
+  output:
+    qc_adapted_work_premium_couple_amount: 4_523.2
+
+
+- name: Quebec adapted work premium amount for couple with children with family income above the turndown threshold
+  period: 2022
+  input:
+    qc_adapted_work_premium_eligibility: true
+    is_married: true
+    count_children: 1
+    adjusted_family_net_income: 30_000
+    province_code: QC
+  output:
+    qc_adapted_work_premium_couple_amount: 3_904.8 # 4523.2 - 0.1*(30000-23816)
+
+- name: Quebec adapted work premium amount for couple with children with family income above the maximum income
+  period: 2022
+  input:
+    qc_adapted_work_premium_eligibility: true
+    is_married: true
+    count_children: 1
+    adjusted_family_net_income: 70_000
+    province_code: QC
+  output:
+    qc_adapted_work_premium_couple_amount: 0 # max(0, 4523.2 - 0.1*(70000-23816))
+
+- name: Quebec adapted work premium amount for couple without child with family income below the turndown threshold 
+  period: 2022
+  input:
+    qc_adapted_work_premium_eligibility: true
+    is_married: true
+    count_children: 0
+    adjusted_family_net_income: 11_000 
+    province_code: QC
+  output:
+    qc_adapted_work_premium_couple_amount: 3_075.78
+
+
+- name: Quebec adapted work premium amount for couple without child with family income above the turndown threshold
+  period: 2022
+  input:
+    qc_adapted_work_premium_eligibility: true
+    is_married: true
+    count_children: 0
+    adjusted_family_net_income: 30_000
+    province_code: QC
+  output:
+    qc_adapted_work_premium_couple_amount: 2_457.38 # 3075.78 - 0.1*(30000-23816)
+
+- name: Quebec adapted work premium amount for couple without child with family income above the maximum income 
+  period: 2022
+  input:
+    qc_adapted_work_premium_eligibility: true
+    is_married: true
+    count_children: 0
+    adjusted_family_net_income: 60_000
+    province_code: QC
+  output:
+    qc_adapted_work_premium_couple_amount: 0 # max(0, 3075.78 - 0.1*(60000-23816))
