+++ policyengine_canada/tests/gov/provinces/qc/tax/income/credits/work_premium/adapted_work_premium/qc_adapted_work_premium_eligibility.yaml
@@ -1,0 +1,124 @@
+- name: Quebec adapted work premium eligibility for household with no disabled members
+  period: 2022
+  input:
+    people:
+      head:
+        is_disabled: false
+        is_head_or_spouse: true
+        working_income: 2_000
+      spouse:
+        is_disabled: false
+        is_head_or_spouse: true
+        working_income: 2_000
+    household:
+      members: [head, spouse]
+      qc_work_premium_eligibility: true
+      province_code: QC
+  output:
+    qc_adapted_work_premium_eligibility: false
+
+- name: Quebec adapted work premium eligibility for married household with disabled head and working income above the requirement
+  period: 2022
+  input:
+    people:
+      head:
+        is_disabled: true
+        is_head_or_spouse: true
+        working_income: 2_000
+      spouse:
+        is_disabled: false
+        is_head_or_spouse: true
+        working_income: 2_000
+    household:
+      members: [head, spouse]
+      qc_work_premium_eligibility: true
+      province_code: QC
+  output:
+    qc_adapted_work_premium_eligibility: true
+
+- name: Quebec adapted work premium eligibility for married household with disabled head but working income below the requirement
+  period: 2022
+  input:
+    people:
+      head:
+        is_disabled: true
+        is_head_or_spouse: true
+        working_income: 1_000
+      spouse:
+        is_disabled: false
+        is_head_or_spouse: true
+        working_income: 2_000
+    household:
+      members: [head, spouse]
+      qc_work_premium_eligibility: true
+      province_code: QC
+  output:
+    qc_adapted_work_premium_eligibility: false
+
+- name: Quebec adapted work premium eligibility for married household with disabled spouse and working income above the requirement
+  period: 2022
+  input:
+    people:
+      head:
+        is_disabled: false
+        is_head_or_spouse: true
+        working_income: 2_000
+      spouse:
+        is_disabled: true
+        is_head_or_spouse: true
+        working_income: 2_000
+    household:
+      members: [head, spouse]
+      qc_work_premium_eligibility: true
+      province_code: QC
+  output:
+    qc_adapted_work_premium_eligibility: true
+
+- name: Quebec adapted work premium eligibility for unmarried household with disabled head and working income above the requirement
+  period: 2022
+  input:
+    people:
+      head:
+        is_disabled: true
+        is_head_or_spouse: true
+        working_income: 2_000
+    household:
+      members: [head]
+      qc_work_premium_eligibility: true
+      province_code: QC
+  output:
+    qc_adapted_work_premium_eligibility: true
+
+- name: Quebec adapted work premium eligibility for unmarried household with non-eligible disabled head and working income above the requirement
+  period: 2022
+  input:
+    people:
+      head:
+        is_disabled: true
+        is_head_or_spouse: true
+        working_income: 2_000
+    household:
+      members: [head]
+      qc_work_premium_eligibility: false
+      province_code: QC
+  output:
+    qc_adapted_work_premium_eligibility: false
+
+- name: Quebec adapted work premium eligibility for household with non-disabled head but disabled other members
+  period: 2022
+  input:
+    people:
+      head:
+        is_disabled: true
+        is_head_or_spouse: true
+        working_income: 2_000
+      dependant:
+        is_disabled: true
+        is_head_or_spouse: false
+        working_income: 2_000
+    household:
+      members: [head]
+      qc_work_premium_eligibility: false
+      province_code: QC
+  output:
+    qc_adapted_work_premium_eligibility: false
