+++ policyengine_canada/tests/gov/provinces/qc/tax/income/credits/work_premium/adapted_work_premium/qc_adapted_work_premium_single_amount.yaml
@@ -1,0 +1,65 @@
+- name: Quebec adapted work premium amount for single parent with family income below the turndown threshold
+  period: 2022
+  input:
+    qc_adapted_work_premium_eligibility: true
+    is_married: false
+    count_children: 1
+    adjusted_family_net_income: 11_000 
+    province_code: QC
+  output:
+    qc_adapted_work_premium_single_amount: 3_628.5
+
+- name: Quebec adapted work premium amount for single parent with family income above the turndown threshold
+  period: 2022
+  input:
+    qc_adapted_work_premium_eligibility: true
+    is_married: false
+    count_children: 1
+    adjusted_family_net_income: 30_000
+    province_code: QC
+  output:
+    qc_adapted_work_premium_single_amount: 2_199.9 # 3628.50 - 0.1*(30000-15714)
+
+- name: Quebec adapted work premium amount for single parent with family income above the maximum income
+  period: 2022
+  input:
+    qc_adapted_work_premium_eligibility: true
+    is_married: false
+    count_children: 1
+    adjusted_family_net_income: 52_000
+    province_code: QC
+  output:
+    qc_adapted_work_premium_single_amount: 0 # max(0, 3628.50 - 0.1*(52000-15714))
+
+- name: Quebec adapted work premium amount for person living alone with family income below the turndown threshold
+  period: 2022
+  input:
+    qc_adapted_work_premium_eligibility: true
+    is_married: false
+    count_children: 0
+    adjusted_family_net_income: 11_000 
+    province_code: QC
+  output:
+    qc_adapted_work_premium_single_amount: 1_973.9
+
+- name: Quebec adapted work premium amount for person living alone with family income above the turndown threshold
+  period: 2022
+  input:
+    qc_adapted_work_premium_eligibility: true
+    is_married: false
+    count_children: 0
+    adjusted_family_net_income: 20_000
+    province_code: QC
+  output:
+    qc_adapted_work_premium_single_amount: 1_545.3 # 1973.9 - 0.1*(20000-15714)
+
+- name: Quebec adapted work premium amount for person living alone with family income above the maximum income
+  period: 2022
+  input:
+    qc_adapted_work_premium_eligibility: true
+    is_married: false
+    count_children: 0
+    adjusted_family_net_income: 52_000
+    province_code: QC
+  output:
+    qc_adapted_work_premium_single_amount: 0 # max(0, 1973.9 - 0.1*(52000-15714))
