+++ policyengine_canada/tests/gov/provinces/qc/tax/income/credits/work_premium/general_work_premium/qc_work_premium_couple_amount.yaml
