--- policyengine_canada/tests/gov/provinces/qc/tax/income/credits/work_premium/qc_adapted_work_premium_eligibility.yaml
@@ -1,70 +1,0 @@
-- name: Quebec adapted work premium eligibility for household with no disabled members
-  period: 2022
-  input:
-    people:
-      head:
-        is_disabled: false
-      spouse:
-        is_disabled: false
-    household:
-      members: [head, spouse]
-      qc_work_premium_eligibility: true
-      province_code: QC
-  output:
-    qc_adapted_work_premium_eligibility: false
-
-- name: Quebec adapted work premium eligibility for married household with disabled head
-  period: 2022
-  input:
-    people:
-      head:
-        is_disabled: true
-      spouse:
-        is_disabled: false
-    household:
-      members: [head, spouse]
-      qc_work_premium_eligibility: true
-      province_code: QC
-  output:
-    qc_adapted_work_premium_eligibility: true
-
-- name: Quebec adapted work premium eligibility for married household with disabled spouse
-  period: 2022
-  input:
-    people:
-      head:
-        is_disabled: false
-      spouse:
-        is_disabled: true
-    household:
-      members: [head, spouse]
-      qc_work_premium_eligibility: true
-      province_code: QC
-  output:
-    qc_adapted_work_premium_eligibility: true
-
-- name: Quebec adapted work premium eligibility for unmarried household with disabled head
-  period: 2022
-  input:
-    people:
-      head:
-        is_disabled: true
-    household:
-      members: [head]
-      qc_work_premium_eligibility: true
-      province_code: QC
-  output:
-    qc_adapted_work_premium_eligibility: true
-
-- name: Quebec adapted work premium eligibility for unmarried household with non-eligible disabled head
-  period: 2022
-  input:
-    people:
-      head:
-        is_disabled: true
-    household:
-      members: [head]
-      qc_work_premium_eligibility: false
-      province_code: QC
-  output:
-    qc_adapted_work_premium_eligibility: false
