+++ policyengine_canada/variables/gov/provinces/qc/tax/income/credits/work_premium/general_work_premium/qc_work_premium_eligibility.py
@@ -4,7 +4,7 @@ from policyengine_canada.model_api import *
 class qc_work_premium_eligibility(Variable):
     value_type = bool
     entity = Household
-    label = "Quebec work premium tax credit requirements"
+    label = "Eligible household for the Quebec work premium tax credit"
     definition_period = YEAR
     defined_for = ProvinceCode.QC
